+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 6.19.4
+
+* added function returning details for a request
+* updated referrer blacklist
+* updated dependencies
+
 ## 6.19.3
 
 * added `Equal` to filter